+++ tasks/tests/tasks_tests/test_collect_tasks/dupe/dupe.py
@@ -1,0 +1,11 @@
+from robocorp.tasks import task
+
+
+@task
+def main():
+    pass
+
+
+@task  # type: ignore # noqa
+def main():  # type: ignore # noqa
+    pass